Manage to Upload the Stats to the backend server based on the protocol. The program is running on the GLinux and Arduino. Refactor the file structure.
